Split to one test file per fixture.
